requirements.txt
pytorch_lightning==1.0.8
numpy==1.19.4
#torch==1.7.0
pandas==1.1.4
mmh3==2.5.1
regex==2020.9.27
rich==1.1.3
click==7.1.2
scikit_learn==0.23.2
datasets==1.1.3
loguru==0.5.3
